Fix up Python interpreter selection logic (#6)
Updated the interpreter logic to check if python3 exists in the PATH
before falling back to just python. This should work on more distros
that don't supply a python binary anymore and just python3.

Also improves the interpreterPath logic to use the provided path exactly
as is for the command.